moneta_veritas/requirements.txt
asgiref==3.10.0
Django==3.2.16
django_debug_toolbar==3.8.1
pytz==2025.2
sqlparse==0.5.3
